requirements.txt
streamlit==1.28.1
plotly==5.17.0
psycopg2-binary==2.9.7
pandas==2.1.1
numpy==1.26.4
altair==5.5.0